more patience while deleting

+++ pyLibrary/env/elasticsearch.py
@@ -25,7 +25,7 @@ from pyLibrary.strings import utf82unicode
 from pyLibrary.dot import coalesce, Null, Dict
 from pyLibrary.dot.lists import DictList
 from pyLibrary.dot import wrap, unwrap
-from pyLibrary.thread.threads import ThreadedQueue
+from pyLibrary.thread.threads import ThreadedQueue, Thread
 
 
 class Index(object):
@@ -770,18 +770,29 @@ class Alias(object):
         if self.debug:
             Log.note("Delete bugs:\n{{query}}", {"query": query})
 
-        result = self.cluster.delete(
-            self.path + "/_query",
-            data=convert.value2json(query),
-            timeout=60
-        )
+        keep_trying = True
+        while keep_trying:
+            result = self.cluster.delete(
+                self.path + "/_query",
+                data=convert.value2json(query),
+                timeout=60
+            )
+            keep_trying = False
+            for name, status in result._indices.items():
+                if status._shards.failed > 0:
+                    if status._shards.failures[0].reason.find("rejected execution (queue capacity ") >= 0:
+                        keep_trying = True
+                        Thread.sleep(seconds=5)
+                        break
 
-        for name, status in result._indices.items():
-            if status._shards.failed > 0:
-                Log.error("ES shard(s) report Failure to delete from {{index}}.  Query was {{query}}", {
-                    "index": name,
-                    "query": query
-                })
+            if not keep_trying:
+                for name, status in result._indices.items():
+                    if status._shards.failed > 0:
+                        Log.error("ES shard(s) report Failure to delete from {{index}}.  Query was {{query}}", {
+                            "index": name,
+                            "query": query
+                        })
+
 
     def search(self, query, timeout=None):
         query = wrap(query)
